+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - new configuration option `implicit_validation` +
 - new validation methods `Essence::validateAll`, `Essence::validate` (alias of `Essence::go`)
-- new configuration methods `Essence::addLink` and `Essence::addMatcher`
+- new configuration methods `Essence::addLink` and `Essence::addMatcher` +
 - wrapper `expect` now works with other data types (not only `Essence` instances) +
 - `AbstractMatcher` now will automatically replace matcher's `value` property value